--- mvc/src/test/java/samples/TestInterfaceController.java
@@ -1,13 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class TestInterfaceController implements Controller {
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        return "viewName";
-    }
-}
